2차 중간발표 자료

+++ BulletinBoard/src/main/java/com/example/demo/domain/Category.java
@@ -1,11 +1,13 @@
 package com.example.demo.domain;
 
 import javax.persistence.Entity;
+import javax.persistence.FetchType;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
+import javax.persistence.OneToMany;
 
 import java.util.*;
 import lombok.Getter;
@@ -19,6 +21,9 @@ import lombok.ToString;
 public class Category {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long CategoryCode;//PK
+	private Integer CategoryCode;//PK
 	private String CategoryName;
+	
+//	@OneToMany(mappedBy="category", fetch=FetchType.EAGER)
+//	private List<SBoard> sboardList = new ArrayList<SBoard>();
 }
