+++ BulletinBoard/src/main/java/com/example/demo/persistence/CategoryRepository.java
@@ -4,5 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.example.demo.domain.Category;
 
-public interface CategoryRepository extends CrudRepository<Category, Long> {
+public interface CategoryRepository extends CrudRepository<Category, Integer> {
+	//Category findByCategoryName(String search);
 }
