+++ BulletinBoard/src/main/java/com/example/demo/persistence/MajorCourseRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.example.demo.domain.MajorCourse;
 
-public interface MajorCourseRepository extends CrudRepository<MajorCourse, Long> {
+public interface MajorCourseRepository extends CrudRepository<MajorCourse, Integer> {
 	List<MajorCourse> findByMajor(String searchword);
 }
